// src/apis/model/viewNotificationPaginatedResponse.ts
/**
 * Generated by orval v6.14.3 🍺
 * Do not edit manually.
 * Farmer App API
 * OpenAPI spec version: v1
 */
import type { ViewNotificationResponse } from './viewNotificationResponse';

export interface ViewNotificationPaginatedResponse {
  totalCount?: number;
  viewNotificationResponses?: ViewNotificationResponse[] | null;
}