// src/apis/model/viewNotificationResponse.ts
/**
 * Generated by orval v6.14.3 🍺
 * Do not edit manually.
 * Farmer App API
 * OpenAPI spec version: v1
 */
import type { NotificationType } from './notificationType';

export interface ViewNotificationResponse {
  id?: string;
  notificationId?: string;
  userId?: string | null;
  isRead?: boolean;
  messageTitle?: string | null;
  messageBody?: string | null;
  notificationType?: NotificationType;
  isProcessed?: boolean;
  msgCreatedBy?: string | null;
  msgCreatedOn?: string;
  msgUpdatedBy?: string | null;
  msgUpdatedOn?: string | null;
  viewMsgCreatedBy?: string | null;
  viewMsgCreatedOn?: string;
  viewMsgUpdatedBy?: string | null;
  viewMsgUpdatedOn?: string | null;
}